requirements.txt
fastapi>=0.68.0,<0.89.1
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
email-validator
python-multipart

#hashing
passlib[bcrypt]>= 1.7.4
#jwt
python-jose[cryptography]